abhishek has run into a crasher here
terraform-provider-clouddns appears to be crashing on the original line
with the following error:
interface conversion: interface {} is nil, not string
So this attempts to check that the field is not nil before converting
it. Further more returning the error is incorrectly attempting to grab
the first result in the array and return it when the array is empty.